move: 패키지 디렉토리 수정
