+++ src/main/java/localgaji/albbaim/__core__/auth/AuthUserArgumentResolver.java
@@ -1,10 +1,9 @@
-package localgaji.albbaim.__core__;
+package localgaji.albbaim.__core__.auth;
 
 import localgaji.albbaim.__core__.exception.CustomException;
 import localgaji.albbaim.__core__.exception.ErrorType;
 import localgaji.albbaim.auth.user.User;
 import localgaji.albbaim.auth.user.UserRepository;
-import localgaji.albbaim.auth.token.TokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
